Silence nonsense compiler warning about shifts wider than type
I had assumed the optimizer would replace the switch with the single case that applied since the `sizeof` operator is evaluated at compile time, but it seems that with gcc 4.8.4 on Ubuntu this isn't the case (maybe the warning is evaluated before the optimizer runs). Since this macro was only used a few times, I simply expanded it.