Do not treat truncated attribute as fatal error
It seems that sometimes the length of an attribute conflicts with the reported total path attributes length. In this case we emit a warning and skip the truncated attributed. This is a case specifically described in Section 4 of RFC 7606 (https://tools.ietf.org/html/rfc7606#section-4) and should be considered as "treat-as-withdraw". We may want to consider in-band signaling of this type of error so that users can handle a corrupt message that doesn't impact the ability of the parser to continue parsing (as is the case with the current INVALID_MSG error code).