Fix bug when iterating through path attributes
We were previously reading path attributes while `nread` was less than the total length of the path attributes data. The problem here is that `nread` is set to 2 at the start of iteration due to reading the `uint16_t` length field. The length field does not include itself in the length value, so halting iteration at `nread == path_attrs->len` was incorrect.

Thanks to @ojasgupta for helping validate this fix.