Give up parsing when reported MRT length does not match data
Previously we were trying to continue parsing when the length given in the MRT header did not match the actual data we parsed, but this turns out to be the wrong choice since it leads to feeding garbage into the parser. Now we simply error out and stop parsing. In the examples that I have seen that trigger this code block, none could have been saved by carrying on parsing anyway, so it isn't like we're losing anything by taking this approach.