Default to not copying BGP marker field into result structure
Since the marker field bits MUST all always be set to 1, for normal parsing this field has no use, so avoiding the memcpy from the buffer into the result structure improves parsing efficiency. Users that want to inspect the contents of this field can set the `marker_copy` option to a non-zero value.