Allow truncated bgp updates to be used, not just skipped.
In particular, NLRIs (announcements/withdrawals) up to the truncation
point are not lost.

Implementation:
If the MRT decoder tells the BGP decoder that truncation is allowed
(because it knows the BGP message is too large to fit in the MRT
message), then the BGP decoder can return PARSEBGP_TRUNCATED_MSG instead
of PARSEBGP_PARTIAL_MSG and ensures that the returned data is not
corrupt.