Replace an assert with an invalid message error
We found some messages from the new FRR codebase that seem to have trailing content included in the BMP message, but that libparsebgp does not parse (it seems that two BGP messages are included in the BMP message). We don't know the cause of this unusual message yet, but libparsebgp should not seg fault in any case.